Make changes to the parser & formatter
Parser changes
* correctly handle separator after the address
This makes parsing a single letter encoded tag kind work.

Formatter
encode tag kinds which are using ascii as a single letter kind, rather
than a field.

The only missing step to parse and format tags generated by ctags is to
preserve kinds which we do not recognize.

TODO: the parser could be simplified.