combineTags & combineTagsPipe
Better check for the same path.  Using module path rather than
'ml_hs_file'.  In case of pre-processed files the latter will point to
the created `.hs` file, while the latter shows a path of the original
file, e.g. `.hsc`.  The drawback is that this file path is relative to
the project's cabal file, which might not be the same directory in which
`ghc-tags-plugin` operates.  To go around this, `combineTags` checks
if a tag is from file which ends with the passed filename.