Non streaming updating of ctags
By default do not use streaming, just read the tags file update it with
the found tags and write back the result to the tags file.  The
`--stream` (`-s`) option is provided to turn on streaming.  Which might
be a bit faster but is also an experimental feature, which might be
removed in the future.